requirements.txt
anyio==4.0.0
APScheduler==3.10.4
boto3==1.29.3
botocore==1.32.3
certifi==2023.11.17
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
jmespath==1.0.1
python-dateutil==2.8.2
python-telegram-bot==20.6
pytz==2023.3.post1
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
tzlocal==5.2
urllib3==2.0.7
